requirements.txt
flask
python-dotenv
requests
werkzeug

# Flask==2.3.2
# requests==2.31.0
# werkzeug==2.3.6